# Orchestrator/static/openapi.json
openapi: 3.0.0
info:
  version: 1.0.0
  title: CAELUS Digital Twin
  description: An interface to the Digital Twin simulation package running on some cluster.

components:
  schemas:
    DroneConfig:
      type: object
      properties:
        mass:
          type: number
          format: float
          example: 1.2
        max_thrust:
          type: number
          format: float
          example: 4
        max_torque:
          type: number
          format: float
          example: 0.05
        arm_length:
          type: number
          format: float
          example: 0.30
    LatLonArray:
      type: array
      items:
        type: number
        format: float
      example:
        [
        -4.346594476,
        55.856459795,
        34.135934166412675
        ]
    UUID4:
      type: string
      example: "d40a18aa-6151-4593-8d30-44cdd3e97d36"
    MissionPayload:
      type: object
      properties:
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/LatLonArray"
        operation_id:
          $ref: "#/components/schemas/UUID4"
        control_area_id:
          $ref: "#/components/schemas/UUID4"
        operation_reference_number:
          $ref: "#/components/schemas/UUID4"
        drone_id:
          $ref: "#/components/schemas/UUID4"
        drone_registration_number:
          $ref: "#/components/schemas/UUID4"
        dis_auth_token:
          $ref: "#/components/schemas/UUID4"
        thermal_model_timestep:
          type: number
          format: float
          example: 0.004
        aeroacoustic_model_timestep:
          type: number
          format: float
          example: 0.004
        drone_config:
          $ref: "#/components/schemas/DroneConfig"
        g_acceleration:
          type: number
          format: float
          example: 9.81
        effective_start_time:
          type: number
          format: integer
          example: 1638526876
        initial_lon_lat_alt:
          $ref: "#/components/schemas/LatLonArray"
        final_lon_lat_alt:
          $ref: "#/components/schemas/LatLonArray"
    
paths:
  /login:
    post:
      tags:
        - Authentication
      description: Login for an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login -- JWT token is provided along with expiry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJ0eXAiOiJ...oQHfTjaqiT01JUpk_y0C46Mns
                  expires:
                    type: string
                    example: Wed, 01 Dec 2021 17:48:50 GMT
        '401':
          description: Authentication failed -- Wrong credentials / User not registered.
          
  /register:
    post:
      tags:
        - Authentication
      description: Register a new user. Only the user with username "admin" can use this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful registration. Proceed with login.
        '400':
          description: Registration failed. An user with the same username already exists.
  /new_mission:
    post:
      tags:
        - Simulation
      description: Create a new mission for the logged in account.
      requestBody:
        required: true
        description: The body of this request should contain the JSON file with the information required by the simulator.
        content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionPayload'
                
      responses:
        '200':
          description: Successful job creation
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    example: eyJ0eXAiOiJ...oQHfTjaqiT01JUpk_y0C46Mns
        '401':
          description: Not logged in
        '402':
          description: Requested docker image hasn't been found.
          
  /simulation_data/{operation_id}:
    get:
      tags:
        - Simulation
      description: Retrieves the simulation data for a given operation
      parameters:
        - in: path
          name: operation_id
          schema:
            type: string
            example: 5c17c4d2-e878-4a47-a6a3-724f16f46ba2
          required: true
          description: UUID of the operation to retrieve information about.
          
      responses:
        '200':
          description: Successfully retrieved simulation data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  battery_level:
                    type: array
                    items:
                      type: number
                      format: float
                      example: [100, 99.9, 99.8]
                  payload_temperature:
                    type: array
                    items:
                      type: number
                      format: float
                      example: [5.0, 5.001, 5.002]
                  aeroacoustic:
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        format: float
                      example: [202751.4487, 635565.9294, 0, 14.58, 0.0, -0.00010678921514848534, 0.0005663992152687102, 0.9999998338939824, 0.0, -0.00010678921514848534, 0.0005663992152687102, 0.9999998338939824, 0.0, -0.00010678921514848534, 0.0005663992152687102, 0.9999998338939824, 0.0, -0.00010678921514848534, 0.0005663992152687102, 0.9999998338939824]
                  flight_time:
                    type: number
                    format: float
                    example: 419.636
                    description: in seconds
        '401':
          description: Authentication failed -- Wrong credentials / User not registered.
        '404':
          description: There is a job associated with the given operation_id but no simulation data is present. This could mean that the simulation is still running.
  /jobs:
    get:
      tags:
        - Simulation
      description: Gets all jobs associated with a user

      responses:
        '200':
          description: Successful login -- JWT token is provided along with expiry.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: eyJ0eXAiOiJ...oQHfTjaqiT01JUpk_y0C46Mns
                    docker_image:
                      type: string
                      example: 'caelus_dt_0.0.1'
                    mission_payload:
                      $ref: "#/components/schemas/MissionPayload"
                    status:
                      type: number
                      enum: [0, 1, 2, 3, 4]
                    status_str:
                      type: string
                      enum: [CREATED, RUNNING, ERROR, TERMINATED, HALTED]
                    issuer_id:
                      type: string
                      example: 'admin'
        '401':
          description: Authentication failed -- Wrong credentials / User not registered.
  /halt/{pid}:
    post:
      tags:
        - Simulation
      parameters:
        - in: path
          name: pid
          schema:
            type: string
            example: 5c17c4d2-e878-4a47-a6a3-724f16f46ba2
          required: true
          description: UUID of the mission to halt
      description: Halt a mission prematurely.
      responses:
        '200':
          description: Successful termination.
        '401':
          description: Authentication failed -- Wrong credentials / User not registered / You do not own that mission.
        '400':
          description: The mission could not be terminated (Internal error or the mission is already in a terminal state)
          